Change type and logic of pv_filesystem_path_get_size to get a better error handling

+++ utils/filesystem.h
@@ -33,7 +33,7 @@ void pv_filesystem_path_concat(char *buf, int size, ...);
 int pv_filesystem_mkdir_p(const char *path, mode_t mode);
 int pv_filesystem_path_remove(const char *path, bool recursive);
 int pv_filesystem_path_rename(const char *src_path, const char *dst_path);
-size_t pv_filesystem_path_get_size(const char *path);
+off_t pv_filesystem_path_get_size(const char *path);
 int pv_filesystem_file_tmp(char *tmp, const char *fname);
 int pv_filesystem_file_save(const char *fname, const char *data, mode_t mode);
 int pv_filesystem_file_copy_from_path(const char *src, const char *dst, mode_t mode);
